# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notebook Api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/notebooks": {
            "get": {
                "tags": [
                    "Notebooks"
                ],
                "summary": "Get list of notebooks",
                "description": "Get list of notebooks",
                "operationId": "570b5cac9e8538fba335224a976bb0da",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "surname": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "patronymic": {
                                                        "type": "string"
                                                    },
                                                    "campaign": {
                                                        "type": "string"
                                                    },
                                                    "phone": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "date_of_birth": {
                                                        "type": "string"
                                                    },
                                                    "image": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "type": "notebooks",
                                                    "id": "1",
                                                    "attributes": {
                                                        "surname": "Иванов",
                                                        "name": "Иван",
                                                        "patronymic": "Иванович",
                                                        "campaign": "Газпром",
                                                        "phone": "89600710772",
                                                        "email": "qwe@mail.ru",
                                                        "date_of_birth": "29.11.1996",
                                                        "image": "/public/images"
                                                    }
                                                },
                                                {
                                                    "type": "notebooks",
                                                    "id": "2",
                                                    "attributes": {
                                                        "surname": "Петров",
                                                        "name": "Петр",
                                                        "patronymic": "Петрович",
                                                        "campaign": "Лукойл",
                                                        "phone": "89600710772",
                                                        "email": "asd@mail.ru",
                                                        "date_of_birth": "29.11.2000",
                                                        "image": "/public/images"
                                                    }
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebooks/store"
                ],
                "summary": "Store notebooks in DB",
                "description": "Store notebooks in DB",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "surname",
                                    "name",
                                    "patronymic",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "surname": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Иванов"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Иван"
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Иванович"
                                    },
                                    "campaign": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Лукойл"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "89600710772"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "qwe@mail.ru"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "29.11.1996"
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/notebooks/{id}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Show notebook",
                "description": "Show notebook",
                "operationId": "03d932ac7745dac1bf7a94bc72e3abf8",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "type": "notebook",
                                                "id": "1",
                                                "attributes": {
                                                    "surname": "Иванов",
                                                    "name": "Иван",
                                                    "patronymic": "Иванович",
                                                    "campaign": "Газпром",
                                                    "phone": "89600710772",
                                                    "email": "qwe@mail.ru",
                                                    "date_of_birth": "29.11.1996",
                                                    "image": "/public/images"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebooks/update"
                ],
                "summary": "Update notebook in DB",
                "description": "Update notebook in DB",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Notebook",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "surname",
                                    "name",
                                    "patronymic",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "surname": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Иванов"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Иван"
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Иванович"
                                    },
                                    "campaign": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Лукойл"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "89600710772"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "qwe@mail.ru"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "29.11.1996"
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "api/v1/notebooks/{id}": {
            "delete": {
                "tags": [
                    "Notebooks/delete"
                ],
                "summary": "Delete Notebook",
                "description": "Delete Notebook",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Notebook",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Success"
                    }
                }
            }
        }
    }
}